Remove leftover regions and unused usings
